Sketch new lazy-thread class
The idea is to support the basic operations of (a) suspending, and (b)
requiring someone else to come out of suspension. These are independent of
other pieces of the runtime, including weight resolution. So to force a
weight, you will have to do two things: register with the weight promise, and
then force the associated thread.

I named the forcing operation "acquire/release" so there can me multiple
concurrent forcers.